Tests pass.

+++ scattermoe/triton_implementation/kernels.py
@@ -52,11 +52,11 @@ def scatter2scatter_triton_kernel(
     acc = tl.zeros((BLOCK_M, BLOCK_N), dtype=ACC_TYPE)
     E_first_idx = tl.min(E_idxs)
     E_last_idx = tl.minimum(tl.max(E_idxs), E - 1)
-    M_idx = tl.load(grouped_idx_ptr + M_block, mask=M_boundary_mask, other=0)
+    M_idx = tl.load(grouped_idx_ptr + M_block, mask=M_boundary_mask, other=0).to(tl.int32)
 
     for E_idx in range(E_first_idx, E_last_idx + 1):
         E_mask = E_idxs == E_idx
-        E_M_idx = tl.where(E_mask, M_idx, 0).to(tl.int32)
+        E_M_idx = tl.where(E_mask, M_idx, 0)
         if x_grouped:
             M_in_idx = M_block
         else:
